+++ model/sequence_cells.jl
@@ -89,7 +89,7 @@ function build_output_table(cellGenomes, mutInductionTimes, driver_muts, group, 
     
     raw_vafs, seed_origins = mutationFrequency(cellGenomes, pop_size, seed_origin)
     
-    output_table = DataFrame(mutation = Int[], raw_vaf = Float64[], vaf = Float64[], t= Int[], is_driver = Int[], seed_parent = Int[], group = String[])
+    output_table = DataFrame(mutation = Int[], raw_vaf = Float64[], vaf = Float64[], t= Float64[], is_driver = Int[], seed_parent = Int[], group = String[])
     
     for (mutation, frequency) in raw_vafs
         t = get(mutInductionTimes, mutation, -1) # -1 if can't find the mutation induction time in dict
